# BackEnd/tshirt-store-api/week-2/openapi.yaml
openapi: 3.0.3
info:
  title: T-Shirt Store API
  description: |
    REST API for the T-Shirt Store capstone project.
    Supports three roles: Manager, Client, and Delivery Person.
  version: 1.0.0
  contact:
    name: Adrian
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Reusable parameters ────────────────────────────────────────
...
paths:
...
  /webhooks/stripe:
    post:
      tags: [Webhooks]
      summary: Handle Stripe webhook events
      description: |
        Handles `payment_intent.succeeded` and `checkout.session.completed` events.
        Verifies the webhook signature. Updates order status and decrements stock.
      operationId: handleStripeWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook processed
        "400":
          description: Invalid signature

  # ── Notifications ────────────────────────────────────────────────
...